Resolve utils include issue
